move nbcparser code to node-bootstrapper
