+++ node-bootstrapper/parser/templates/containerd.toml.gtpl
