+++ node-bootstrapper/parser/templates/cse_cmd.sh.gtpl
