+++ node-bootstrapper/parser/templates/kubenet-cni.json.gtpl
