+++ node-bootstrapper/parser/testdata/test_nbc.json
